move taggable concern to model concerns
